#906: Move jvm options logic to super class IdeaBasedIdeToolCommandlet to work for every ide-based tool and add method getIdeProductPrefix
