+++ cli/src/main/java/com/devonfw/tools/ide/tool/androidstudio/AndroidStudio.java
@@ -52,6 +52,12 @@ public class AndroidStudio extends IdeaBasedIdeToolCommandlet {
   }
 
   @Override
+  protected String getIdeProductPrefix() {
+
+    return STUDIO;
+  }
+
+  @Override
   public void setEnvironment(EnvironmentContext environmentContext, ToolInstallation toolInstallation, boolean additionalInstallation) {
 
     super.setEnvironment(environmentContext, toolInstallation, additionalInstallation);
